Updated password settings, desktop environment
The password is now stored in a separate file, password.txt, which is
not version-controlled. This is for security reasons.

The desktop environment was switched from KDE to Gnome. KDE is
consistently crashing and making my windows inaccessible (permanently
minimized), so I'm going to use Gnome until I have more time and more
experience with Nix.