Reorganized directory on advice of nix-dev emails
I was trying to organize my configuration based on the attribute being
set; that is, there would be one file that lists all the packages, one
file that lists all the services, etc. Mutliple people on the nix-dev
mailing list suggested that I use the imports feature to bring files
into the configuration in an additive way; this allows my to have, for
example, one file that deals with my terminal configuration and makes
sure that terminal-related programs are installed, while another file
deals with the GUI and also lists programs to be installed. Importing
both of these files merges the list of installed programs, along with
any programs that the base configuration.nix file lists. This is a more
felxible solution and encourages more practical modules.